+++ chaosk8s/probes.py
@@ -1,11 +1,13 @@
 # -*- coding: utf-8 -*-
+from datetime import datetime
 import json
 import os.path
-from typing import Union
+from typing import Dict, Union
 import urllib3
 
 from chaoslib.exceptions import FailedActivity
 from chaoslib.types import MicroservicesStatus, Secrets
+import dateparser
 from logzero import logger
 from kubernetes import client, watch
 import yaml
@@ -15,7 +17,7 @@ from chaosk8s import create_k8s_api_client
 
 __all__ = ["all_microservices_healthy", "microservice_available_and_healthy",
            "microservice_is_not_available", "service_endpoint_is_initialized",
-           "deployment_is_not_fully_available"]
+           "deployment_is_not_fully_available", "read_microservices_logs"]
 
 
 def all_microservices_healthy(ns: str = "default",
@@ -176,3 +178,50 @@ def deployment_is_not_fully_available(name: str, ns: str= "default",
         raise FailedActivity(
             "microservice '{name}' failed to stop running within {t}s".format(
                 name=name, t=timeout))
+
+
+def read_microservices_logs(name: str, since_seconds: Union[str, None] = None,
+                            ns: str="default", with_previous: bool=False,
+                            secrets: Secrets=None) -> Dict[str, str]:
+    """
+    Fetch logs for all the pods with the label `"name"` set to `name` and
+    return a dictionnary with the keys being the pod's name and the values
+    the logs of said pod.
+
+    If you provide `since_seconds`, this returns the logs of the last N seconds
+    until now. This can set to a fluent delta such as `10 minutes`.
+
+    You may also set `with_previous` to `True` to capture the logs of a
+    previous pod's incarnation, if any.
+    """
+    api = create_k8s_api_client(secrets)
+    v1 = client.CoreV1Api(api)
+    ret = v1.list_namespaced_pod(
+        ns, label_selector="name  in ({name})".format(name=name))
+
+    logger.debug("Found {d} pods: [{p}]".format(
+        d=len(ret.items), p=', '.join([p.metadata.name for p in ret.items])))
+
+    since = since_seconds
+    if since_seconds:
+        now = datetime.now()
+        since = int((now - dateparser.parse(since_seconds)).total_seconds())
+
+    params = dict(
+        namespace=ns,
+        follow=False,
+        timestamps=True,
+        _preload_content=False
+    )
+
+    if since:
+        params["since_seconds"] = since
+
+    logs = {}
+    for p in ret.items:
+        name = p.metadata.name
+        logger.debug("Fetching logs for pod '{n}'".format(n=name))
+        r = v1.read_namespaced_pod_log(name, **params)
+        logs[name] = r.read().decode('utf-8')
+
+    return logs